aurora.back/requirements.txt
asgiref==3.4.1
attrs==21.4.0
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.1
django-braces==1.15.0
django-cors-headers==3.10.1
django-filter==21.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-mysql==4.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
factory-boy==3.2.1
Faker==11.3.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
model-bakery==1.4.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
